Task 6
Added a new module square_generator.py
