+++ PPY_Task_03/main.py
@@ -1,6 +1,6 @@
-#Task 1
 import math
-
+import square_generator as sg
+#Task 1
 list1 = [i ** 2 for i in range(1, 11)]
 
 print(list1)
@@ -16,22 +16,22 @@ print(list_square_creator(1, 11))
 
 #Task 3
 
-class SquareGenerator:
-    def list_square_creating(start, end):
+# class SquareGenerator:
+#     def list_square_creating(start, end):
+#
+#         #Task 5
+#         if start <= end:
+#             return [i**2 for i in range(start, end)]
+#         else:
+#             print('Your start is greater than end!')
+#             return []
 
-        #Task 5
-        if start <= end:
-            return [i**2 for i in range(start, end)]
-        else:
-            print('Your start is greater than end!')
-            return []
-
-print(SquareGenerator.list_square_creating(1, 11))
+print(sg.SquareGenerator.list_square_creating(1, 11))
 
 
 #Task 4
 
-list4 = SquareGenerator.list_square_creating(1, 11)
+list4 = sg.SquareGenerator.list_square_creating(1, 11)
 for x in range(len(list4)):
     list4[x] = math.sqrt(list4[x])
 print(list4)
